+++ modules/audio-playback/src/loader.js
@@ -39,7 +39,7 @@ function Loader(ctx, events) {
 module.exports = Loader;
 
 function fetchLocalOrRemote(url) {
-  if (false && process.env.NODE_ENV !== "production") {
+  if (process.env.NODE_ENV !== "production") {
     const local = url.replace(STORAGE, "http://localhost:3333/data/audiosets");
     return fetch(local).catch(function() {
       return fetch(url);