+++ modules/audioset/src/migrate.js
@@ -1,4 +1,5 @@
 import hsvToRgb from "hsv-rgb";
+import { resourceUrl } from "@atpls/audioset";
 
 const HSV = /^hsv\((.*)\)$/;
 function rgb(color) {
@@ -11,21 +12,14 @@ function rgb(color) {
   return "rgb(" + c[0] + "," + c[1] + "," + c[2] + ")";
 }
 
-function resourceUrl(name, resource) {
-  const url =
-    resource && resource.length
-      ? resource[0].replace("{{filename}}", name)
-      : "";
-  return url;
-}
-
 export default function migrate(set) {
-  console.log("migrate", set.id);
+  console.log("joder", set.loader.sources);
   Object.keys(set.clips).forEach(function(name) {
     const clip = set.clips[name];
     clip.id = name;
     clip.display.color = rgb(clip.display.color);
     clip.display.cover = resourceUrl(name, set.loader.sources.covers);
+    clip.display.audiosrc = resourceUrl(name, set.loader.sources.audio);
     if (set.audio.defaults) Object.assign(clip.audio, set.audio.defaults);
   });
   return set;