+++ modules/server/src/static.js
@@ -3,7 +3,7 @@ const express = require("express");
 // Static files
 const CONFIG = {
   "/data": "../../../data",
-  "/audio": "../../../data/audiosets",
+  "/audiosets": "../../../data/audiosets",
   "/sets": "../../audioset/sets",
   "/app": "../../app-visuals/build",
   "/control": "../../app-control/build",
@@ -15,6 +15,7 @@ const router = express.Router();
 
 Object.keys(CONFIG).forEach(url => {
   const path = join(__dirname, CONFIG[url]);
+  console.log("joder", url, path);
   router.use(url, express.static(path));
 });